refactor: database lambda에서 screeshot 레이어 역할 분리
